requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
loguru>=0.7.0 
torch==2.5.1
transformers>=4.30.0
datasets==3.2.0
pandas==2.2.3
sentencepiece==0.2.0
